changed around stuff in spotsReducer and made some update-spot progress
